+++ src/uniforms/buffer.rs
@@ -1,5 +1,4 @@
 use buffer::{BufferView, BufferViewAny, BufferType, BufferCreationError};
-use buffer::Mapping as BufferMapping;
 use uniforms::{AsUniformValue, UniformValue, UniformBlock, UniformType};
 
 use std::ops::{Deref, DerefMut};
@@ -10,11 +9,6 @@ use backend::Facade;
 #[derive(Debug)]
 pub struct UniformBuffer<T> where T: Copy {
     buffer: BufferView<T>,
-}
-
-/// Mapping of a buffer in memory.
-pub struct Mapping<'a, T> {
-    mapping: BufferMapping<'a, T>,
 }
 
 /// Same as `UniformBuffer` but doesn't contain any information about the type.
@@ -36,7 +30,7 @@ impl<T> UniformBuffer<T> where T: Copy {
 
     /// Uploads data in the uniforms buffer.
     pub fn new_if_supported<F>(facade: &F, data: T) -> Option<UniformBuffer<T>> where F: Facade {
-        let buffer = match BufferView::new(facade, &[data], BufferType::UniformBuffer, true) {
+        let buffer = match BufferView::new(facade, &data, BufferType::UniformBuffer, true) {
             Ok(b) => b,
             Err(BufferCreationError::BufferTypeNotSupported) => return None,
             e @ Err(_) => e.unwrap(),
@@ -59,7 +53,7 @@ impl<T> UniformBuffer<T> where T: Copy {
 
     /// Creates an empty buffer.
     pub fn empty_if_supported<F>(facade: &F) -> Option<UniformBuffer<T>> where F: Facade {
-        let buffer = match BufferView::empty(facade, BufferType::UniformBuffer, 1, true) {
+        let buffer = match BufferView::empty(facade, BufferType::UniformBuffer, true) {
             Ok(b) => b,
             Err(BufferCreationError::BufferTypeNotSupported) => return None,
             e @ Err(_) => e.unwrap(),
@@ -69,31 +63,6 @@ impl<T> UniformBuffer<T> where T: Copy {
             buffer: buffer,
         })
     }
-
-    /// Modifies the content of the buffer.
-    pub fn upload(&mut self, data: T) {
-        self.write(&[data]);
-    }
-
-    /// Reads the content of the buffer if supported.
-    pub fn read_if_supported(&self) -> Option<T> {
-        self.buffer.read_if_supported().and_then(|buf| buf.into_iter().next())
-    }
-
-    /// Reads the content of the buffer.
-    ///
-    /// # Features
-    ///
-    /// Only available if the 'gl_read_buffer' feature is enabled.
-    #[cfg(feature = "gl_read_buffer")]
-    pub fn read(&self) -> T {
-        self.read_if_supported().unwrap()
-    }
-
-    /// Maps the buffer in memory.
-    pub fn map(&mut self) -> Mapping<T> {
-        Mapping { mapping: self.buffer.map() }
-    }
 }
 
 impl<T> Deref for UniformBuffer<T> where T: Copy {
@@ -101,20 +70,6 @@ impl<T> Deref for UniformBuffer<T> where T: Copy {
 
     fn deref(&self) -> &BufferView<T> {
         &self.buffer
-    }
-}
-
-impl<'a, D> Deref for Mapping<'a, D> {
-    type Target = D;
-
-    fn deref(&self) -> &D {
-        &self.mapping.deref()[0]
-    }
-}
-
-impl<'a, D> DerefMut for Mapping<'a, D> {
-    fn deref_mut(&mut self) -> &mut D {
-        &mut self.mapping.deref_mut()[0]
     }
 }